Add github action renovate helper (#282)
By default, if you use a github action with a digest (git commit) renovate will keep the digest inline with the HEAD of the detault branch - so you may well be ahead of the latest release. It will also ignore the comment next to the digest..

Code snipets like user/action@digest # v1.2.3 will actually be follwing the head default branch and not the released version.

Adding this configuration setting means renovate will look at the comment and use that to find the digest.